test: moved testing.py to tests/test_signin.py
I updated the Makefile to conform to github/jidn/python-Makefile
and part of that was creating a tests directory.  To make pytest
find the test file it needed to match test_*.py

Additionally jsontest.com was kicking out 503 for over quota. The test
is fixed to fail gracefully when that happens.  I wonder what the quota
period is for that service?  It was the first of the month when it
happened.